refactor: moved Library/methods to utils/license
