--- src/data/Library/index.ts
@@ -1,1 +1,0 @@
-export * from './methods'
